Convert action to modules
