+++ action/post.mjs
@@ -1,21 +1,21 @@
-const path = require('node:path')
-const fs = require("node:fs/promises");
-const { getState, DECAY_PID_KEY, TEMP_DIR } = require("./util");
+import path from "node:path";
+import fs from "node:fs/promises";
+import { getState, DECAY_PID_KEY, TEMP_DIR } from "./util.mjs";
 
 const pid = getState(DECAY_PID_KEY);
 
-if (typeof pid === 'undefined') {
+if (typeof pid === "undefined") {
   console.error(`${DECAY_PID_KEY} state could not be found`);
   process.exit(1);
 }
 
 // @TODO: Check whether the server is actually running
-console.log(`Decay server will be stopped on pid: ${pid}`)
+console.log(`Decay server will be stopped on pid: ${pid}`);
 process.kill(parseInt(pid));
 
 function noop(error) {
-  console.error(error)
-  return ""
+  console.error(error);
+  return "error reading output file";
 }
 
 Promise.all([
@@ -25,10 +25,8 @@ Promise.all([
   if (std) {
     console.log(`Server output: `, std);
   }
-  
+
   if (error) {
-    console.error(`Server errors: `, err)
+    console.error(`Server errors: `, err);
   }
 });
-
-
